requirements.txt
python-telegram-bot==13.7
requests==2.28.2
python-dotenv==0.21.0
redis==4.5.4
psycopg2-binary==2.9.6
